requirements.txt
motor==3.6.0
python-telegram-bot~=21.6
pyyaml ~=6.0